+++ x11/fresh.cu
@@ -77,7 +77,7 @@ extern "C" int scanhash_fresh(int thr_id, uint32_t *pdata,
 	uint32_t endiandata[20];
 
 	int throughput = opt_work_size ? opt_work_size : (1 << 19); // 256*256*8;
-	throughput = min(throughput, max_nonce - first_nonce);
+	throughput = min(throughput, (int) (max_nonce - first_nonce));
 
 	if (opt_benchmark)
 		((uint32_t*)ptarget)[7] = 0x00ff;
